feature: code-refactoring refactored base validator
